core: use doubles everywhere input is concerned
The current logic relies on integers to pass input coordinates. This is
bad for accuracy, not to mention the cases where scaling is enabled,
because input is in surface-local coordinates.

+++ plugins/single_plugins/fisheye.cpp
@@ -168,7 +168,7 @@ class wayfire_fisheye : public wf::plugin_interface_t
             const wf_framebuffer_base& dest)
         {
             auto oc = output->get_cursor_position();
-            wlr_box box = {oc.x, oc.y, 1, 1};
+            wlr_box box = {(int)oc.x, (int)oc.y, 1, 1};
             box = output->render->get_target_framebuffer().
                 framebuffer_box_from_geometry_box(box);
             oc.x = box.x;
